pyproject.toml
[tool.poetry]
name = "basketball_app"
version = "0.1.0"
description = ""
authors = ["Nita Alexandru <allex_eu@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.6"
django-rest-framework = "^0.1.0"
django-custom-user = "^1.1"
djangorestframework-simplejwt = "^5.2.2"
requests = "^2.28.2"
drf-yasg = "^1.21.5"
python-dotenv = "^0.21.1"
pytest = "^7.2.1"
pytest-django = "^4.5.2"
black = "^23.1.0"
flake8 = "^6.0.0"
isort = "^5.12.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
